Pick percentile items consistently in histograms
The slot_slide version of histograms picks percentile items from the skew
heap (if configured). In this case, it picks from a reversed list, since
the heap keeps only the largest values. A slight difference in percent
calculations could cause an off-by-one error when comparing the items
picked by the exometer_slide and the exometer_slot_slide histograms.