Fix reporter crashes when a cache is used for function entries
Previously `exometer_reporter` process crashed in case a cache is
used for function entries. The reason is a wrong processing of the
result `{error, unavailable}` from `exometer_function:get_value/4`.
There is no supervision tree for the related processes under
`exometer_core` so the reporter processes will also fail which is
bad for stateful reporters.

This commit fixes the processing for `{error, unavailable}` and the
behavior of function entries with and without cache should be the same.

Note that this commit does not change a returned value of
`exometer:get_value/2` for such entries.
It's still '{ok, {error, unavailable}}'.